add missing pieces to growth participant
