+++ packages/graphql/src/typeDefs/Growth.js
@@ -154,7 +154,7 @@ module.exports =
     # Sends email invites with referral code on behalf of the referrer.
     invite(emails: [String!]!): Boolean
     # Enrolls user into the growth engine program.
-    enroll(accountId: ID!, agreementMessage: String!, signature: String!, inviteCode: String): EnrollResponse
+    enroll(accountId: ID!, agreementMessage: String!, signature: String!, inviteCode: String, fingerprint: String): EnrollResponse
     # Records a growth engine event.
     log(event: JSON!): Boolean
     # Remind a user that his invitation is still pending
